chore(errors): move concrete errors into parent-named subpackages
RuleViolatedError → rule_violations/ (subclass of RuleViolationError)
ValidationFailedError → validation/ (subclass of ValidationError)

Update all consumer imports in validators, field_errors, __init__.py,
and tests to reflect the new subpackage paths.
