+++ src/forging_blocks/foundation/errors/rule_violations/__init__.py
@@ -1,0 +1,3 @@
+from .rule_violated_error import RuleViolatedError
+
+__all__ = ["RuleViolatedError"]
