+++ src/forging_blocks/foundation/errors/rule_violations/rule_violated_error.py
