+++ src/forging_blocks/foundation/errors/validation/__init__.py
@@ -1,0 +1,3 @@
+from .validation_failed_error import ValidationFailedError
+
+__all__ = ["ValidationFailedError"]
