+++ src/forging_blocks/foundation/errors/validation/validation_failed_error.py
